Move ZuoraOrdersApiPrimitives to Libraries
