+++ lambda/src/test/scala/pricemigrationengine/libs/ZuoraOrdersApiPrimitivesTest.scala
@@ -1,9 +1,5 @@
-package pricemigrationengine.migrations
+package pricemigrationengine.libs
 
-import pricemigrationengine.model._
-
-import java.time.LocalDate
-import pricemigrationengine.model.ZuoraOrdersApiPrimitives
 import ujson.Value
 
 class ZuoraOrdersAPIPrimitivesTest extends munit.FunSuite {
